+++ src/main/java/de/tum/in/www1/artemis/config/SecurityConfiguration.java
@@ -165,7 +165,7 @@ public class SecurityConfiguration extends WebSecurityConfigurerAdapter {
             .antMatchers("/api/lti/launch/*").permitAll()
             .antMatchers("/api/files/attachments/lecture/**").permitAll()
             .antMatchers("/api/files/attachments/attachment-unit/**").permitAll()
-            .antMatchers("/api/files/attachments/exercise/*/build-plan").permitAll()
+            .antMatchers("/api/programming-exercises/*/build-plan").permitAll()
             .antMatchers("/api/files/file-upload-exercises/**").permitAll()
             .antMatchers("/api/files/markdown/**").permitAll()
             .antMatchers("/api/**").authenticated()
